refactor: rename grade_scroll to grade_scroll_position and fix argument order
